pyproject.toml
[tool.poetry]
name = "mandle"
version = "0.1.0"
description = ""
authors = ["arjoonn sharma <arjoonn.94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
numpy = "^1.18.4"
scikit-image = "^0.17.2"
pillow = "^7.1.2"
bottle = "^0.12.18"
bottle_tools = "^2019.12.22"
tqdm = "^4.46.0"
matplotlib = "^3.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
